requirements.txt
attrs==18.1.0
click==6.7
more-itertools==4.3.0
numpy==1.15.0
pandas==0.23.3
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
pytest==3.5.0
python-dateutil==2.7.3
python-dotenv==0.9.0
pytz==2018.5
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.10
